+++ src/main/config/config.c
@@ -27,38 +27,38 @@
 #include "common/axis.h"
 #include "common/maths.h"
 
-#include "flight/flight.h"
-
 #include "drivers/sensor.h"
 #include "drivers/accgyro.h"
 #include "drivers/compass.h"
-
 #include "drivers/system.h"
 #include "drivers/gpio.h"
 #include "drivers/timer.h"
 #include "drivers/pwm_rx.h"
+#include "drivers/serial.h"
 
 #include "sensors/sensors.h"
 #include "sensors/gyro.h"
 #include "sensors/compass.h"
-
-#include "io/statusindicator.h"
 #include "sensors/acceleration.h"
 #include "sensors/barometer.h"
-#include "drivers/serial.h"
-#include "io/serial.h"
-#include "telemetry/telemetry.h"
-
-#include "flight/mixer.h"
 #include "sensors/boardalignment.h"
 #include "sensors/battery.h"
+
+#include "io/statusindicator.h"
+#include "io/serial.h"
 #include "io/gimbal.h"
 #include "io/escservo.h"
-#include "rx/rx.h"
 #include "io/rc_controls.h"
 #include "io/rc_curves.h"
 #include "io/ledstrip.h"
 #include "io/gps.h"
+
+#include "rx/rx.h"
+
+#include "telemetry/telemetry.h"
+
+#include "flight/mixer.h"
+#include "flight/flight.h"
 #include "flight/failsafe.h"
 #include "flight/altitudehold.h"
 #include "flight/imu.h"
