+++ src/main/flight/flight.h
@@ -65,43 +65,8 @@ typedef enum {
 
 #define FLIGHT_DYNAMICS_INDEX_COUNT 3
 
-typedef struct int16_flightDynamicsTrims_s {
-    int16_t roll;
-    int16_t pitch;
-    int16_t yaw;
-} flightDynamicsTrims_def_t;
-
-typedef union {
-    int16_t raw[3];
-    flightDynamicsTrims_def_t values;
-} flightDynamicsTrims_t;
-
-typedef struct rollAndPitchTrims_s {
-    int16_t roll;
-    int16_t pitch;
-} rollAndPitchTrims_t_def;
-
-typedef union {
-    int16_t raw[2];
-    rollAndPitchTrims_t_def values;
-} rollAndPitchTrims_t;
-
-typedef struct rollAndPitchInclination_s {
-    // absolute angle inclination in multiple of 0.1 degree    180 deg = 1800
-    int16_t rollDeciDegrees;
-    int16_t pitchDeciDegrees;
-} rollAndPitchInclination_t_def;
-
-typedef union {
-    int16_t raw[ANGLE_INDEX_COUNT];
-    rollAndPitchInclination_t_def values;
-} rollAndPitchInclination_t;
-
-
 #define DEGREES_TO_DECIDEGREES(angle) (angle * 10)
 #define DECIDEGREES_TO_DEGREES(angle) (angle / 10.0f)
-
-extern rollAndPitchInclination_t inclination;
 
 extern int16_t gyroData[FLIGHT_DYNAMICS_INDEX_COUNT];
 extern int16_t gyroZero[FLIGHT_DYNAMICS_INDEX_COUNT];
@@ -118,7 +83,6 @@ extern int32_t AltHold;
 extern int32_t vario;
 
 void setPIDController(int type);
-void resetRollAndPitchTrims(rollAndPitchTrims_t *rollAndPitchTrims);
 void resetErrorAngle(void);
 void resetErrorGyro(void);
