+++ src/main/io/rc_controls.c
@@ -31,9 +31,6 @@
 
 #include "drivers/system.h"
 
-#include "flight/flight.h"
-#include "flight/navigation.h"
-
 #include "drivers/sensor.h"
 #include "drivers/accgyro.h"
 
@@ -45,12 +42,17 @@
 
 #include "io/gps.h"
 #include "io/beeper.h"
-#include "mw.h"
 
 #include "rx/rx.h"
 #include "io/escservo.h"
 #include "io/rc_controls.h"
 #include "io/rc_curves.h"
+
+#include "flight/flight.h"
+#include "flight/navigation.h"
+
+#include "mw.h"
+
 
 static escAndServoConfig_t *escAndServoConfig;
 static pidProfile_t *pidProfile;
