+++ src/main/sensors/sensors.h
@@ -17,6 +17,18 @@
 
 #pragma once
 
+
+typedef struct int16_flightDynamicsTrims_s {
+    int16_t roll;
+    int16_t pitch;
+    int16_t yaw;
+} flightDynamicsTrims_def_t;
+
+typedef union {
+    int16_t raw[3];
+    flightDynamicsTrims_def_t values;
+} flightDynamicsTrims_t;
+
 #define CALIBRATING_GYRO_CYCLES             1000
 #define CALIBRATING_ACC_CYCLES              400
 #define CALIBRATING_BARO_CYCLES             200 // 10 seconds init_delay + 200 * 25 ms = 15 seconds before ground pressure settles
